refactor(apps): the screen fixture gets its own address
`scriptedScreenAssembler` lands `app.tsx` through `authoredScreen`; the
wire fixture beside it compiles `app.vendo` through `authored`. Two
different seams sharing one module, and the wire one is on its way out —
so the screen half moves to `testing/screen-assembler.ts` and
`authoring-assembler.ts` goes back to exactly what it was.
